Rework ABI compatibility
- Build wheels using a "manylinux3" (CentOS 7-based) deps image instead of manylinux1

   - The official manylinux1 image from quay.io/pypa/manylinux1_x86_64, which
     is based on CentOS 5 (with the addition of the RedHat devtoolset-2 from
     CentOS 6) results in built DALI binaries that segfault when importing
     alongside pre-built PyTorch.  This appears to be a C++11 ABI incompatibility
     that manifested in the PyTorch DALI plugin due to pybind11 and that was
     unresolved until CentOS 7's devtoolset-3.  We created a minimally altered
     "manylinux3" derivative of manylinux1 so as to allow building DALI
     binaries that would be compatible with pre-built binaries for all of
     TensorFlow, MXNet, and PyTorch and now use this as a default.  The deps
     image's base can be changed back to the upstream manylinux1 using a
     command like this:
     ```
     docker build -t dali:deps --build-arg FROM_IMAGE_NAME=quay.io/pypa/manylinux1_x86_64 -f Dockerfile.deps .
     ```
     Fixes https://github.com/NVIDIA/DALI/issues/13 .

- Statically link most dependencies, *excluding* libstdc++

   - This doesn't significantly inflate the package size, and it gives us much more
     control over what symbols will eventually be exported into the application (which
     might have other OpenCV or Protobuf [etc] versions already dynamically loaded)

   - This has the fringe benefit of statically linking in the small parts of NPP 9.0
     that we depended on, so now DALI binaries can work with applications built with
     other CUDA versions >= 9.0 without us necessarily needing to have different
     binaries for each CUDA.

   - Didn't statically link turbojpeg at this time because it was causes a symbol
     clash with the libjpeg that's bundled in to OpenCV.  That said, libturbojpeg
     uses symbol versioning, so this should be okay.

   - Didn't statically link libstdc++, because having multiple sets of various C++11
     standard library static items in the process space leads to segfaults.  That
     said, we build with a version of libstdc++ that should provide reasonable
     symbol compatibility, and symbols are versioned.

- Create a version map for ld

   - This lets us map all of the symbols that do get exported from libdali.so
     (including from protobuf) and tag them as specific to DALI; this should avoid
     a scenario whereby multiple protobufs in the process space could otherwise
     conflict with each other.

- Fix an RPATH bug in bundle-wheel.sh

   - Fixes https://github.com/NVIDIA/DALI/issues/5 , wherein the RPATH that
     would correctly point to libdali-*.so was not being set in backend_impl*.so
     or libdali_tf.so subsequent to the addition of an `nvidia` namespace
     on the `nvidia.dali` package just prior to the release of DALI 0.1.0.

+++ dali/python/bundle-wheel.sh
@@ -88,24 +88,12 @@ make_wheel_record() {
 }
 
 DEPS_LIST=(
-    "/usr/local/lib/liblmdb.so"
     "dali/libdali.so"
-    "/usr/local/lib/libnvjpeg.so.9.0"
-    "/usr/local/lib/libopencv_core.so.3.1"
-    "/usr/local/lib/libopencv_imgcodecs.so.3.1"
-    "/usr/local/lib/libopencv_imgproc.so.3.1"
-    "/usr/local/lib/libprotobuf.so.15"
     "/usr/local/lib/libturbojpeg.so.0"
 )
 
 DEPS_SONAME=(
-    "liblmdb.so"
     "libdali.so"
-    "libnvjpeg.so.9.0"
-    "libopencv_core.so.3.1"
-    "libopencv_imgcodecs.so.3.1"
-    "libopencv_imgproc.so.3.1"
-    "libprotobuf.so.15"
     "libturbojpeg.so.0"
 )
 
@@ -121,9 +109,14 @@ for filepath in "${DEPS_LIST[@]}"; do
     filename=$(basename $filepath)
     patchedname=$(fname_with_sha256 $filepath)
     patchedpath=$PKGNAME_PATH/.libs/$patchedname
+    patched+=("$patchedname")
+
+    if [[ ! -f "$filepath" ]]; then
+        echo "Didn't find $filename, skipping..."
+        continue
+    fi
     echo "Copying $filepath to $patchedpath"
     cp $filepath $TMPDIR/$patchedpath
-    patched+=("$patchedname")
 
     # HACK: CUDA libraries like libnvjpeg.so.9.0 set their .gnu.version info
     # in the ELF... to their own name.  which means that if we go messing
@@ -162,17 +155,11 @@ find $PKGNAME_PATH -name '*.so*' -o -name '*.bin' | while read sofile; do
 done
 
 # set RPATH of backend_impl.so and similar to $ORIGIN, $ORIGIN/.libs
-find $PKGNAME_PATH -maxdepth 1 -type f -name "*.so*" | while read sofile; do
-    echo "Setting rpath of $sofile to " '$ORIGIN:$ORIGIN/.libs'
-    patchelf --set-rpath '$ORIGIN:$ORIGIN/.libs' $sofile
-    patchelf --print-rpath $sofile
-done
-
-# set RPATH of test/*.bin to $ORIGIN, $ORIGIN/.libs
-find $PKGNAME_PATH/test -maxdepth 1 -type f -name "*.bin" | while read sofile; do
-    echo "Setting rpath of $sofile to " '$ORIGIN:$ORIGIN/.libs'
-    patchelf --set-rpath '$ORIGIN:$ORIGIN/../.libs' $sofile
-    patchelf --print-rpath $sofile
+find $PKGNAME_PATH/* -type f -name "*.so*" -o -name "*.bin" | while read FILE; do
+    UPDIRS=$(dirname $(echo "$FILE" | sed "s|$PKGNAME_PATH||") | sed 's/[^\/][^\/]*/../g')
+    echo "Setting rpath of $FILE to '\$ORIGIN:\$ORIGIN$UPDIRS/.libs'"
+    patchelf --set-rpath "\$ORIGIN:\$ORIGIN$UPDIRS/.libs" $FILE
+    patchelf --print-rpath $FILE
 done
 
 # set RPATH of .libs/ files to $ORIGIN
